+++ src/pipelines/graphql/mutations.generated.tsx
@@ -24,6 +24,13 @@ export type CreatePipelineFromTemplateVersionMutationVariables = Types.Exact<{
 
 export type CreatePipelineFromTemplateVersionMutation = { __typename?: 'Mutation', createPipelineFromTemplateVersion: { __typename?: 'CreatePipelineFromTemplateVersionResult', success: boolean, errors?: Array<Types.CreatePipelineFromTemplateVersionError> | null, pipeline?: { __typename?: 'Pipeline', id: string, name?: string | null, code: string } | null } };
 
+export type UpgradePipelineVersionFromTemplateMutationVariables = Types.Exact<{
+  input: Types.UpgradePipelineVersionFromTemplateInput;
+}>;
+
+
+export type UpgradePipelineVersionFromTemplateMutation = { __typename?: 'Mutation', upgradePipelineVersionFromTemplate: { __typename?: 'UpgradePipelineVersionFromTemplateResult', success: boolean, errors: Array<Types.UpgradePipelineVersionFromTemplateError> } };
+
 
 export const UpdatePipelineDocument = gql`
     mutation UpdatePipeline($input: UpdateDAGInput!) {
@@ -145,3 +152,37 @@ export function useCreatePipelineFromTemplateVersionMutation(baseOptions?: Apoll
 export type CreatePipelineFromTemplateVersionMutationHookResult = ReturnType<typeof useCreatePipelineFromTemplateVersionMutation>;
 export type CreatePipelineFromTemplateVersionMutationResult = Apollo.MutationResult<CreatePipelineFromTemplateVersionMutation>;
 export type CreatePipelineFromTemplateVersionMutationOptions = Apollo.BaseMutationOptions<CreatePipelineFromTemplateVersionMutation, CreatePipelineFromTemplateVersionMutationVariables>;
+export const UpgradePipelineVersionFromTemplateDocument = gql`
+    mutation upgradePipelineVersionFromTemplate($input: UpgradePipelineVersionFromTemplateInput!) {
+  upgradePipelineVersionFromTemplate(input: $input) {
+    success
+    errors
+  }
+}
+    `;
+export type UpgradePipelineVersionFromTemplateMutationFn = Apollo.MutationFunction<UpgradePipelineVersionFromTemplateMutation, UpgradePipelineVersionFromTemplateMutationVariables>;
+
+/**
+ * __useUpgradePipelineVersionFromTemplateMutation__
+ *
+ * To run a mutation, you first call `useUpgradePipelineVersionFromTemplateMutation` within a React component and pass it any options that fit your needs.
+ * When your component renders, `useUpgradePipelineVersionFromTemplateMutation` returns a tuple that includes:
+ * - A mutate function that you can call at any time to execute the mutation
+ * - An object with fields that represent the current status of the mutation's execution
+ *
+ * @param baseOptions options that will be passed into the mutation, supported options are listed on: https://www.apollographql.com/docs/react/api/react-hooks/#options-2;
+ *
+ * @example
+ * const [upgradePipelineVersionFromTemplateMutation, { data, loading, error }] = useUpgradePipelineVersionFromTemplateMutation({
+ *   variables: {
+ *      input: // value for 'input'
+ *   },
+ * });
+ */
+export function useUpgradePipelineVersionFromTemplateMutation(baseOptions?: Apollo.MutationHookOptions<UpgradePipelineVersionFromTemplateMutation, UpgradePipelineVersionFromTemplateMutationVariables>) {
+        const options = {...defaultOptions, ...baseOptions}
+        return Apollo.useMutation<UpgradePipelineVersionFromTemplateMutation, UpgradePipelineVersionFromTemplateMutationVariables>(UpgradePipelineVersionFromTemplateDocument, options);
+      }
+export type UpgradePipelineVersionFromTemplateMutationHookResult = ReturnType<typeof useUpgradePipelineVersionFromTemplateMutation>;
+export type UpgradePipelineVersionFromTemplateMutationResult = Apollo.MutationResult<UpgradePipelineVersionFromTemplateMutation>;
+export type UpgradePipelineVersionFromTemplateMutationOptions = Apollo.BaseMutationOptions<UpgradePipelineVersionFromTemplateMutation, UpgradePipelineVersionFromTemplateMutationVariables>;
